refactor(e2e): extract common PdfDownloader utility from page objects
Removes ~360 lines of duplicated PDF download code across 4 page objects:
- ReceiptCandidatesPage.js
- ShipmentPage.js
- InvoicePage.js
- SalesOrderPage.js

Benefits:
- Single source of truth for PDF download logic
- Uses Playwright request API (fixes axios session bug in CI)
- Consistent behavior across all document types
- Easier to maintain and update

The new PdfDownloader utility handles:
- Opening print modal (Alt+P)
- Both direct download and popup scenarios
- Proper cookie/session handling via Playwright context
- PDF validation (checks %PDF- header)
- Temp file management
